requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
docx2txt==0.8
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
lxml==5.2.1
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pdf2image==1.17.0
pillow==10.3.0
plum-dispatch==1.7.4
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
Spire.Doc==12.4.0
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
